TestSuites: NAND OneNAND: Bad Blocks Handling
'dd' command does not have any code to avoid bad blocks, so as soon as it finds
one, it will exit with flag error enabled. This is not the case for upper
software layers as in jffs2 implementations (which has a block interface to
device and capable of handling bad blocks).

Test case is changed now so dd access block interface once it is mounted
letting bad block handling to jffs2 layer.

Signed-off-by: Abraham Arce <x0066660@ti.com>