TestSuites: Camera: multi_open take device arg
Add an extra cmdline argument to allow the user to set the camera device.

If not given then it will default to 1 (the current hardcoded value).

This change allows the app to be used for other camera devices (like
Sony IMX046).
Also adds change to get current brightness value and reset it after
test completes.

Signed-off-by: Dominic Curran <dcurran@ti.com>