TestSuites: I2C: Deleting lmsensor utility, and disabling Makefile
Due an update in the I2C test, we are going to include i2c-tools instead
lmsensors code. This to made more easy the compilation of the tools needed
for I2C test.

This patch delete all lmsensors code, and disable the Makefile commenting
all lines, to avoid a compilation break.

Signed-off-by: Ricardo Perez Olivares <x0081762@ti.com>