I2C: Fix the failing I2C testcases for OMAP 3430/3630
after upgrading the i2c-tools package.

This patch fixes the the error thrown by I2C testcases in TITAN:

Error: Could not set address to 0x48: Device or resource busy

Recently, the I2C-tools utility package was upgraded to
i2c-tools-3.0.10, this requires a change in parameters
to run the i2c-tools. This patch fixes this issue.

Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>