TestSuites: Camera: Modify scripts to use new frame rate cmdline in burst_mode
Modify scripts to use new frame rate cmdline in burst_mode.
This allows user to specify the frame rate and thus control
which base size will be selected from the sensor.

This is important because before it was using whatever the
last test had used, and thus results were unpredictable.
This led to upscaling from the 720p size, when that wasnt
the original intent of the test.

Because the burstest.sh script changed, then all scenarios
that call this script have had to be modified to pass a
desired frame rate. This means that tests for MT & OV
sensors have been impacted.
All scenarios that have changed have been tested.

Signed-off-by: Dominic Curran <dcurran@ti.com>