TestSuites: Camera: Fix ratio for resize wrapper test
Problem with 'QVGAtoSXGA' resizer wrapper test.
The tests resizer ratio for upscaling was set to high.
4x resizing to not possible according to TRM calculation.

HORZ = (32*sph +(ow-1)*hrsz +16) / 256 + 7

where:
  ow = 1280
  sph = 0
  hrsz = 64 (minimum value allowed)

HORZ = (32*0 +(1280-1)*64 +16) / 256 + 7
     = (0    +1279*64  +16) / 256 + 7
     = 81872 / 256  + 7
     = 319 + 7
HORZ = 326

BUT, our input image is only 320 pixels wide !

This was leading to extra pixels both horz and vertical
getting resized, which caused a green line at bottom
of image.

Fix is to reduce the output size (1248x932) so that resizer
can handle it.

Signed-off-by: Dominic Curran <dcurran@ti.com>