TestSuites: Camera: snapshot format the output filename
Format the output filename so that it is compatible with the
PYUV viewer.

This change means that PYUV can pull all the info about the
image format from the filename and thus no need for user to
input it.

PYUV only displays YUV images.  However, comparable data has
also been added for RAW filename for completeness.

Details of expected filename format for PYUV:
 http://dsplab.diei.unipg.it/~baruffa/dvbt/pyuvhelp/doc2.htm#control

Signed-off-by: Dominic Curran <dcurran@ti.com>