Testsuites:Ethernet:Remove IRQ comparison in test
IRQ numbers for ethernet reported in /proc/interrupts and ifconfig will be
different due to a data mismatch bug in /kernel/net/core/dev.c
This is a known issue that according to the subsystem maintainer should not
be fixed for the sake of not breaking userspace applications.

Hence, this comparison is removed from the testsuite and replaced with a
comparison of the interrupt counter values before and after a ping
operation. The test will fail if there is no increase in the counter.

Signed-off-by: Aharon Campos <aharon.campos@ti.com>