TestSuites: Watchdog: Removing deprecated scenario
This scenario is following a procedure that is not applicable anymore for
L25x. In order to test this functionality the testwdtioctl binary should
be executed then suspend the board using the 'F2' key. After suspending
the system it should be resumed using the 'F1' key and all power domains
should hit target state and the board should be rebooted properly.
This procedure is going to be added in the test case spec and UTR for 25x.

Signed-off-by: Cristopherson Torres <x0096871@ti.com>