TestSuites: Camera: Substitute RAW10 by specific component orders
This should untie RAW10 to always be SGRBG10, which is not always
the case for the available sensor.

The Tester should be more conscious of the pixelformat variant
he's requesting.

The new values will be:
- SGRBG10
- SRGGB10
- SGBRG10
- SBGGR10

Which are the 4 possible combinations for RAW Bayer 10-bit formats,
And should match the newly existing RAW10 pixel formats available
in V4L2 spec.

Signed-off-by: Sergio Aguirre <saaguirre@ti.com>