Testsuites:General:Create Makefile for ddt repo
This patch creates the Makefile for the device driver test repository.
After specifying KDIR, it automatically determines for which version the
testsuites are meant to be compiled, so issuing a make command (with the
proper environment variables exported) should suffice.

Variables TARGET_PLATFORM and TARGET_FILESYSTEM are exported for use of
Makefiles in subdirectories.

During incremental make, only those testsuites that changed are copied, but
this assumes the time stamp of the <suite>/test_code folder has changed
during the execution of the sub-make if anything has been updated.

Additional rules for creating the <suite>/scripts/scenarios folder are
included, to also copy only scenario files that have changed when executing
an incremental make.

Signed-off-by: Aharon Campos <aharon.campos@ti.com>