Makefile:Don't copy text files nor create extra dirs
This fixes the rule for creating the utils directory in the output folder,
to avoid copying text files.

When $(UTILSDIR) was changed from utils to $(PWD)/utils, the entire
directory structure, starting from / was created inside the output folder.
By using the $(notdir) function this issue is solved as well.

Signed-off-by: Aharon Campos <aharon.campos@ti.com>