General: Moving filesystemtype.sh to utils folder
This script was duplicated in different folders and it was the same code.
Moving to a general path where it can be used for any testsuite.
Also, using alias in order to make reference to same 'dmesg' binary when using
different FS (Poky, Busybox, Android).

This patch was taken from 25x branch.

Signed-off-by: Cristopherson Torres <x0096871@ti.com>
Signed-off-by: Aharon Campos <aharon.campos@ti.com>