TestSuites: Camera: snapshot: queue buf with correct userptr
Make sure that when the buffer is queued to the camera that the correct userptr is passed down in the v4l2_buffer.

The v4l2 code kernel that queues the buffer needs this.
If it does not match the pointer that was originally passed
to v4l2 has then the new pointer in the queue request will
be used and a buffer maybe lost track of.

Signed-off-by: Dominic Curran <dcurran@ti.com>