General:Output of filesystemtype.sh in /tmp
Previously, this script used $TMPBASE/tmp to store its output (This would
be the temporary folder for each of the testsuites). However, in 24x the
configuration variables for each testsuite must be different to avoid
conflicts while running smp scenarios.

Using the system's /tmp folder, which will always be there, we avoid the
need to reference all the diferent variables for all testsuites in this
script.

Signed-off-by: Aharon Campos <aharon.campos@ti.com>