Automation: Introducing Test Debug Tool aka TDT
This patch presents the initial support for The TDT

TDT is automation framework based on python which knows
how to fetch a git repository or a specific git branch,
compile the binaries for a specific platform, run a set
of predefined test cases, report the test results, check
each failing test case and bisect the git tree to find
the exact commit object that causes the failure and
report the results.

This tool uses Buildbot as a trigger mechanism for a
daily built-test-debug cycle and TITAN to run the
test cases.

The bisect mechanism is based on git-rev-list and it uses
the following bisection logic to achieve this task:

<http://embetek.com/index.php?option=com_content&view=article&id=56:git-bisect>

The main goal of this tool is:

- Report test results in daily test cycle
- Generates daily binaries for a especific Release/ Revision
- Find the changes that introduced a bug for each test
  cases that failed

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>
Signed-off-by: Cristopherson Torres <x0096871@ti.com>
Signed-off-by: Abraham Arce <x0066660@ti.com>