Makefile: New list of applicable testsuites
When specifying 'all' as the value of testsuites during compilation, the
actual list of testsuites is obtained from the list in the top makefile.

Most of the testsuites apply for all platforms, and only some apply
depending on specific kernel configs.

Based on this, a new list of applicable testsuites is created.

Signed-off-by: Aharon Campos <aharon.campos@ti.com>