Testsuites: Utils: Use if stament instead of test command.
The following command is used in some conf.sh files:
	handlerError.sh "clean"

This command erases a file in /tmp/ folder, if the file does not exist
the script finishes with status 1.
The /tmp/ folder is being cleaned previously by general.configuration
script, so the there is a failure reported during conf.sh execution:
	FATAL: Configuration file with errors

The failure exit status is removed using if staments instead of test command.

Signed-off-by: Cristopherson Torres <x0096871@ti.com>