Testsuites: Ethernet: Network File Sharing
To use this scenario it is needed to have a machine with NFS server running.
The following variables should be exported:
IPADDR:			The computer's IP with NFS server running.
HOST_MOUNT_POINT:	A folder in the host computer where NFS can be mounted.
MOUNT_POINT:		An empty folder in the board FS.

Signed-off-by: Cristopherson Torres <x0096871@ti.com>