Testsuites: handlers: FS type recognition.
Add a new handler which provides capabilities to determine which fs is being
used. This handler has a series of rules based on files structure and if those
rules are satisfied the FS will be identified according to the name asigned.
If none fs is recognized it will return unknown as the FS name.
Initially just recognizes ubuntu and android FS.

Signed-off-by: Cristopherson Torres <x0096871@ti.com>