OMAP4: FB tescase: Corerction in test case
1.In call_test_write function, FBIOGETCMAP gets a the color map settings
and edit it by IOCtl FBIOPUTCMAP.After this change, when it queries the
color map by IOCtl FBIOGETCMAP,it shal be similar to what has been set
through the previous command of FBIOPUTCMAP.The comparison of r, g and
b with cmap.red[0],cmap.green[0] and cmap.blue[0] should yield true in
ideal situation but the coding error caused the test case to fail.

2.Similarly, the return value of the IOCtl FBIOPUT_VSCREENINFO would be zero
if the IOCTL goes through successfully. Here again, the return value was
compared wrongly which made the test case failed.The test case was wrong
in comparison and hence corrected.

Signed-off-by: Sujeet Baranwal <s-baranwal@ti.com>