Adding scripts for resizer HQ
All the srcipt use new yuv images which are provided on the CDP.
Note that the L_DD_CAMERA_0219 is copy of the original L_DD_CAMERA_0119.
It has the same steps but uses the reszhq binary.

Not using resolutions which are not aligned to 32 bytes on width.
Because mem->rsz->mem pipe requires both the input and output images
to be aligned to 32 bytes by width.

Signed-off-by: Ivaylo Petrov <ivpetrov@mm-sol.com>
Signed-off-by: Akwasi Boateng <akwasi.boateng@ti.com>