pmd: add SRF constraint test
New test for exercising the constraint framework.  Loading the module
creates test points in debugfs under a top-level directory named
"constraints".

Within this directory will be dentry's for each type of constraint
available to drivers.  Currently only "min_mpu_freq" is present which
corresponds to omap_pm_set_min_mpu_freq in plat-omap/omap-pm-srf.c.

A constraint can be placed on MPU/VDD1 by writing a frequency value in
kilohertz:
echo 600000000 > /debug/constraints/min_mpu_freq

This can be done as many times as needed to adjust constraints.  Writing
0 or rmmod'ing the module will release all constraints.  Different types
of constraints can be held at a time (once other constraints are
implemented); in fact this is a good way to stress the framework and
test for robustness.

Signed-off-by: Mike Turquette <mturquette@ti.com>
Signed-off-by: Cristopherson Torres <cristopherson.torres@ti.com>