OMAP4: DMA: Reduce number of transfers in each thread
With L24.9 integration test, it was observed that, the memory
allocation used to fail due to no. of elements in the list is set to 15.
The element list is reduced to 2 and no. of threads are increased
from 2 to 4 with this patch.

Code optimization for better handling of threads done.

Signed-off-by: Manjunatha GK <manjugk@ti.com>