Testsuites: Utils: Use OK key to mark as pass.
Change the key code value from U to OK (L24x), ENTER (L27x) and Select (L25x)
to mark a test as pass. The previous key codes are mapped to OK key in
the Zoom3 and OMAP4 boards. Press any other key to mark the test case as fail.

Signed-off-by: Cristopherson Torres <cristopherson.torres@ti.com>