handlerAndroidPM: clear dmesg buffer and add a delay before reading it
Reading the dmesg output continuosly during suspend was not allowing
the Android PM framework to go to suspend since it was throwing the
content of the dmesg buffer all the time, causing the uart to register
the uart_wakelock. Clear the buffer after reading the content and
wait sometime before reading the content again.

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>