Testsuites: Handler: Fix incorrect conditional statement
Move the number of parameters validation inside the if statement. The number
of parameters validation should not be combined with the command validation.
The previous conditional statement was expecting that the command 'run' should
have a number of parameters different than 2. Since 'run' command correct usage
is to have 2 paremeters the condition was never satisfied and a 'operation
introduced is incorrect' message error was displayed.

Signed-off-by: Cristopherson Torres <cristopherson.torres@ti.com>