DMA: Add OSWR test case
Add test case for OSWR testing. The DMA controller will lose
context if CPU0 enters OFF mode. It will not lose context
if it hits mpu retention. This test case will configure
the requested dma channel for data transfer and configuration
register data will be stored in an array. The DMA resources
are not release during DMA callback. Later, the cpu1 is
disabled and CPU0 is suspended for testing OSWR feature.
After wake up, the configuration registers contents are
compared with initial configuration values. If all the register
values are matching, it proves that there is no context
loss with DMA controller registers with core retention.

Tested with latest L27x android kernel image on omap4
SDP with ES2.1 silicon.

Signed-off-by: G, Manjunath Kondaiah <manjugk@ti.com>