Testsuite: Video: Add support for Interlace
Adding the support of interlace in setimg & setwin.

For example:
====================================================
./setimg 1 RGB565 860 520 5
./setwin 1 0 0  1920 1080 5
====================================================

As per V4L2 spec
V4L2_FIELD_SEQ_TB	5
V4L2_FIELD_NONE		1

So "5" signifies that the input buffer is interlaced buffer where as "1" there suggests that it's a progressive buffer.

The imgstream script is modified to take the interlace & time delay parameters.

Signed-off-by: Samreen <samreen@ti.com>