monkey: correct some parameters in the monkey scripts
When a key has been pressed down = action down = 0
When a key has been released = action up = 1

This patch correct the above conditions and also
add set the speed and count values to 10 and 1.0
respectively which are default values for monkey
scripts.

Also instead of going to home screen right after
set the screen timeout value, return to home by
pressing the BACK key several times. This has
shown to be more robust when the script is run
multiple times

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>
Signed-off-by: G, Manjunath Kondaiah <manjugk@ti.com>