Testsuites: Correct L_DD_POWER_0001_0001 test execution and info
L_DD_POWER_0001_0001 is described as following: "Checking that
system time is maintained with dynamic tick during 600 seconds"

Previously the testcase was checking the system time only for 1
minute and it was performing a unnecessary sleep of 10 min before
the test. Also I added a TITAN reboot command before test execution,
this will ensure that the test is not afected by any system
failure that could be induced previously.

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>
Signed-off-by: G, Manjunath Kondaiah <manjugk@ti.com>