Testsuites: i2c: do not use absolute paths for temp file
Do not use absolute paths for any file. You can export
a path if it is inside of the testsuite, but when the
test tries to access or write to a different location,
it can find issues like a non-existing path or permission
denied (ro).

The test_runner uses the tmp folder to write any temporal
file used in a scenario.

This change fixes some issues on the i2c testsuite execution

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>